関数をクラス化

+++ message/message_list_delete.php
@@ -1,23 +1,27 @@
 <?php
 require_once('../config_2.php');
+$user = new User($_POST['user_id']);
+$user = $user->get_user();
 
 try {
 
   $user_id = $_POST['user_id'];
   $destination_user_id = $_POST['destination_user_id'];
   $dbh = db_connect();
-  if (check_relation_delete_message($user_id, $destination_user_id)) {
-    $sql = 'DELETE FROM message_relation WHERE user_id=?';
+  if ($user->check_relation_delete_message($user_id, $destination_user_id)) {
+    $sql = 'DELETE FROM message_relation WHERE user_id=? AND destination_user_id=?';
     $stmt = $dbh->prepare($sql);
     $data[] = $user_id;
+    $data[] = $destination_user_id;
     $stmt->execute($data);
     $dbh = null;
   } else {
-    $sql = 'DELETE FROM message_relation WHERE user_id=?';
+    $sql = 'DELETE FROM message_relation WHERE user_id=? AND destination_user_id=?';
     $stmt = $dbh->prepare($sql);
     $data[] = $user_id;
+    $data[] = $destination_user_id;
     $stmt->execute($data);
-    $sql = 'DELETE FROM message WHERE  (user_id = :user_id and destination_user_id = :destination_user_id) or (user_id = :destination_user_id and destination_user_id = :user_id)';
+    $sql = 'DELETE FROM message WHERE (user_id = :user_id and destination_user_id = :destination_user_id) or (user_id = :destination_user_id and destination_user_id = :user_id)';
     $stmt = $dbh->prepare($sql);
     $stmt->execute(array(
       ':user_id' => $user_id,
@@ -25,10 +29,10 @@ try {
     ));
     $dbh = null;
   }
-
   set_flash('sucsess', 'メッセージリストを削除しました');
   reload();
 } catch (Exception $e) {
+  error_log($e, 3, "../../php/error.log");
   _debug("メッセージ削除失敗");
   exit();
 }
